Add ESLint JSdoc plugin and rules
This update includes the integration of the ESLint JSdoc plugin and the addition of corresponding rules to the eslint.config file. The rules added are aimed at enforcing stylings and conventions that are relevant to documentation within Javascript code. The JSdoc plugin has also been added to the project dependencies in the package.json file.