Update structure of export statements and include files in tsconfig
The updates include modifying the export structure in src/index.ts and specifying the files to be included in tsconfig.json. Changes have also been made to the 'lint:fix' script in package.json ensuring it lints all .ts files and an explicit export statement is added to index.ts.