Update version in deno.json and automate the process
The version in deno.json has been updated. A new command, 'update-version-deno', has been added to the Makefile to extract the version from package.json and update it in deno.json. This command is now part of the 'after:bump' scripts in the release-it configuration file, automating this process for future version updates.