Add retry functionality to API and routes
This update adds retry capabilities to the API and route components of the project. The number of max retries can be configured on both levels. A callback option is implemented to react on retry events, meaning it is invoked whenever a retry is performed. The relevant code logic and test cases are also included in this commit.